requirements.txt
openai
bs4
ipykernel
python-dotenv
requests
duckduckgo_search
readability-lxml
lxml_html_clean
streamlit